[Updated] Refactored MemorizeWords component to utilize a new DisplayVocabularyWord type for better type management. [Enhanced] Improved shuffle functionality to handle both scroll and pagination modes effectively. [Added] Exposed setIsShuffled in useShuffle hook for better state control.
